whd-1134 swap in new CityManager
