+++ ipv-core/src/main/java/com/ibm/whc/deid/resources/Resource.java
@@ -6,9 +6,14 @@
 package com.ibm.whc.deid.resources;
 
 /**
- * 
+ * An item of a particular category, such as a city, gender, religion, name, etc. that can be
+ * uniquely identified within its category by a key.
  */
 public interface Resource {
 
+  /**
+   * 
+   * @return the unique key within its category of items for this item
+   */
   public String getKey();
 }
