+++ ipv-core/src/main/java/com/ibm/whc/deid/util/CityManager.java
@@ -1,5 +1,5 @@
 /*
- * (C) Copyright IBM Corp. 2016,2020
+ * (C) Copyright IBM Corp. 2016,2021
  *
  * SPDX-License-Identifier: Apache-2.0
  */
@@ -7,104 +7,54 @@ package com.ibm.whc.deid.util;
 
 import java.io.IOException;
 import java.io.InputStream;
-import java.security.SecureRandom;
-import java.util.ArrayList;
 import java.util.Collection;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
-
 import org.apache.commons.csv.CSVParser;
 import org.apache.commons.csv.CSVRecord;
-
 import com.ibm.whc.deid.models.City;
-import com.ibm.whc.deid.models.Location;
+import com.ibm.whc.deid.resources.LocalizedResourceManager;
 import com.ibm.whc.deid.shared.localization.Resource;
 import com.ibm.whc.deid.util.localization.LocalizationManager;
 import com.ibm.whc.deid.util.localization.ResourceEntry;
 import com.ibm.whc.deid.utils.log.LogCodes;
+import com.ibm.whc.deid.utils.log.LogManager;
 
-public class CityManager extends ResourceBasedManager<City> {
-  /** */
-  private static final long serialVersionUID = -3274559886741195053L;
+public class CityManager extends LocalizedResourceManager<City> {
 
-  private final SecureRandom random = new SecureRandom();
-
-  private Map<String, List<Location>> cityListMap;
-  private Map<String, LatLonDistance> latLonTree = null;
-
-  private LatLonDistance<City> distanceCalc;
-
-	public CityManager(String tenantId, String localizationProperty) {
-		super(tenantId, Resource.CITY, localizationProperty);
-
-    distanceCalc = new LatLonDistance<City>(getItemList());
+  private static final LogManager logger = LogManager.getInstance();
+    
+  protected CityManager() {
+    // nothing required here
   }
-
-  @Override
-  public Collection<ResourceEntry> getResources() {
-		return LocalizationManager.getInstance(localizationProperty).getResources(Resource.CITY);
-  }
-
-  protected void addToCityList(City city, String countryCode) {
-    List<Location> list = cityListMap.get(countryCode);
-
-    if (list == null) {
-      list = new ArrayList<>();
-      list.add(city);
-      cityListMap.put(countryCode, list);
-    } else {
-      list.add(city);
-    }
-  }
-
-  @Override
-  public Map<String, Map<String, City>> readResourcesFromFile(
-      Collection<ResourceEntry> entries) {
-    Map<String, Map<String, City>> cities = new HashMap<>();
-
-    for (ResourceEntry entry : entries) {
-      InputStream inputStream = entry.createStream();
-      String locale = entry.getCountryCode();
-
-      try (CSVParser reader = Readers.createCSVReaderFromStream(inputStream)) {
-        for (CSVRecord line : reader) {
-          String name = line.get(0);
-          Double latitude = FileUtils.parseDouble(line.get(1));
-          Double longitude = FileUtils.parseDouble(line.get(2));
-          String countryCode = line.get(3);
-          City city = new City(name, latitude, longitude, countryCode, locale);
-
-          addToCityList(city, locale);
-
-          addToMapByLocale(cities, locale, name.toUpperCase(), city);
-          addToMapByLocale(cities, getAllCountriesName(), name.toUpperCase(), city);
+  
+  public static CityManager getCityManager(String localizationProperty) {
+    CityManager cityManager = new CityManager();
+    
+    Collection<ResourceEntry> resourceEntries = LocalizationManager.getInstance(localizationProperty).getResources(Resource.CITY);    
+    for (ResourceEntry entry : resourceEntries) {
+      
+      try (InputStream inputStream = entry.createStream()) {
+        String locale = entry.getCountryCode();    
+        
+        try (CSVParser reader = Readers.createCSVReaderFromStream(inputStream)) {
+          for (CSVRecord line : reader) {
+            String name = line.get(0);
+            Double latitude = FileUtils.parseDouble(line.get(1));
+            Double longitude = FileUtils.parseDouble(line.get(2));
+            String countryCode = line.get(3);
+            City city = new City(name, latitude, longitude, countryCode, locale);
+  
+            String key = city.getKey();
+            cityManager.add(key, city);
+            cityManager.add(locale, key, city);
+          }
         }
-        inputStream.close();
       } catch (IOException | NullPointerException e) {
         logger.logError(LogCodes.WPH1013E, e);
       }
     }
 
-    return cities;
-  }
-
-  @Override
-public void init() {
-    this.cityListMap = new HashMap<>();
-  }
-
-  @Override
-public void postInit() {
-    this.latLonTree = new HashMap<>();
-
-    for (String key : cityListMap.keySet()) {
-      try {
-        this.latLonTree.put(key, new LatLonDistance(cityListMap.get(key)));
-      } catch (Exception e) {
-        logger.logError(LogCodes.WPH1013E, e);
-      }
-    }
+    return cityManager;
   }
 
   /**
@@ -114,22 +64,30 @@ public void postInit() {
    * @param k the k
    * @return the closest city
    */
-  public String getClosestCity(String city, int k) {
-    String key = city.toUpperCase();
-    City lookup = getKey(key);
-
-    if (lookup == null) {
-      return getRandomKey();
+  public String getClosestCity(String cityString, int k) {
+    String value = null;
+    
+    City city = getValue(cityString);
+    
+    if (city == null) {
+      city = getRandomValue();
+      if (city != null) {
+        value = city.getName();
+      }
+      
+    } else {
+      LatLonDistance<City> distanceCalc = new LatLonDistance<>(getValues());
+      List<City> neighbors = distanceCalc.findNearestK(city, k);
+      if (neighbors != null && !neighbors.isEmpty()) {
+        int count = neighbors.size();
+        if (count == 1) {
+          value = neighbors.get(0).getName();
+        } else {
+          value = neighbors.get(random.nextInt(count)).getName();
+        }
+      }
     }
-
-    List<City> neighbors = distanceCalc.findNearestK(lookup, k);
-
-    return (neighbors.get(random.nextInt(neighbors.size()))).getName();
+    
+    return value;
   }
-
-  @Override
-  public Collection<City> getItemList() {
-    return getValues();
-  }
-
 }
